refactor: adjust body parameter
