+++ android/beagle/src/main/java/br/com/zup/beagle/android/networking/HttpAdditionalData.kt
@@ -32,4 +32,4 @@ data class HttpAdditionalData(
     val method: HttpMethod? = HttpMethod.GET,
     val headers: Map<String, String>? = hashMapOf(),
     val body: String? = null,
-): Parcelable
+) : Parcelable
